完善版 修改角色问题

+++ src/main/java/com/sgcc/account/settings/service/MaintainCardServiceImpl.java
@@ -173,8 +173,12 @@ public class MaintainCardServiceImpl implements MaintainCardService {
 
     @Override
     public void grantRole(Map<String, String> param) {
-        // 删除该卡原卡主权限
-        maintainCardDao.deleteRole(param);
+        String uuid = param.get("uuid");
+        if (!"".equals(uuid)){
+            // 删除该卡原卡主权限
+            maintainCardDao.deleteRole(param);
+        }
+
         // 插入新卡主权限
         param.put("pid", CommUtil.getUUID());
         maintainCardDao.addRole(param);
